backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain-anthropic
langchain-core
